Move highlighterCursor code to new folder
