+++ src/contentScripts/api.js
@@ -1,5 +1,5 @@
 import * as highlight from './highlight/index.js';
-import { toggleHighlighterCursor } from './highlighterCursor.js';
+import { toggleHighlighterCursor } from './highlighterCursor/index.js';
 import * as highlights from './highlights/index.js';
 
 const highlighterCursor = {
