+++ src/contentScripts/highlighterCursor/index.js
